Category Delete Functionality Updated and Sweet Alert Added to category and subcategory

+++ app/Http/Controllers/SubCategoryController.php
@@ -7,6 +7,7 @@ use App\SubCategory;
 use Illuminate\Http\Request;
 use Illuminate\Support\Str;
 use Illuminate\Support\Carbon;
+use App\Product;
 
 class SubCategoryController extends Controller
 {
@@ -53,8 +54,14 @@ class SubCategoryController extends Controller
 
     //Sub Category soft delete
     function SubCategoryDelete($id){
-        SubCategory::findOrFail($id)->delete();
-        return back()->with('scategory_delete', 'Sub Category Deleted Successfully!!!');
+        $cat_product = Product::where('subcategory_id', $id)->count();
+
+        if ($cat_product > 0) {
+            return back()->with('ProductAvailable', 'You can not delete the Sub category with existing Product.');
+        } else{
+            SubCategory::findOrFail($id)->delete();
+            return back()->with('scategory_delete', 'Sub Category Deleted Successfully!!!');
+        }
     }
 
     //Sub category Restore
